(mostly) finished
still need to have settings as a separate file, allow a hotkey to re-read them from that file to allow modification to the settings in game. a menu is possible but not really necessary at this stage. triggerbot could be useful as well. (color or distance to aimbot target based)